+++ src/Sentry.Unity/WebGL/SentryWebGL.cs
@@ -17,9 +17,6 @@ public static class SentryWebGL
     {
         options.DiagnosticLogger?.LogDebug("Updating configuration for Unity WebGL.");
 
-        // Due to special exception handling on WebGL we capture these through `LogMessageReceived` too
-        options.AddIntegration(new UnityApplicationLoggingIntegration(captureExceptions: true));
-
         // Note: we need to use a custom background worker which actually doesn't work in the background
         // because Unity doesn't support async (multithreading) yet. This may change in the future so let's watch
         // https://docs.unity3d.com/2019.4/Documentation/ScriptReference/PlayerSettings.WebGL-threadsSupport.html